+++ changelog.md
@@ -1,4 +1,23 @@
 # Changelog
 
+All notable changes to this project will be documented in this file.
 
-# [2.0.0](https://github.com/akki0256/discord-interaction/compare/v1.0.1...v2.0.0) - (2023-01-21)
+## [2.0.0](https://github.com/akki0256/discord-interaction/compare/v1.0.1...v2.0.0) - (2023-01-21)
+
+## [2.0.1](https://github.com/akki0256/discord-interaction/compare/v2.0.0...v2.0.1) - (2023-01-21)
+
+### Typings
+
+**BaseCommand:**
+ - Fix `resetCoolTime()` parameter
+ - Fix `getCoolTime()` parameter
+ - Fix `getLastUseDiff()` parameter
+ - Fix `isInCoolTime()` parameter
+
+**DiscordInteractions:**
+ - Fix `loadInteractions()`return value
+ - Fix `registerCommands()` return value
+
+### Added
+**DiscordInteractions:**
+ -